Combine fetcher files
